+++ shell/agents/AIShell.OpenAI.Agent/AIShell.OpenAI.Agent.csproj
@@ -24,7 +24,7 @@
     <PackageReference Include="Azure.AI.OpenAI" Version="1.0.0-beta.17" />
     <PackageReference Include="Azure.Core" Version="1.39.0" />
     <PackageReference Include="SharpToken" Version="2.0.3" />
-    <PackageReference Include="AIShell.Abstraction" Version="0.1.0-alpha.11">
+    <PackageReference Include="AIShell.Abstraction" Version="0.1.0-alpha.12">
       <ExcludeAssets>contentFiles</ExcludeAssets>
       <PrivateAssets>all</PrivateAssets>
     </PackageReference>
